+++ changelog.md
@@ -2,6 +2,29 @@
 
 > All Time Stamps: GMT+8
 
+## 0.2.29.0 _2020-01-17 15:00_
+
+* 更新Forge依赖版本 update Forge dependency
+* 错误修复 bug fixes
+* 增加部分合成表 added some recipes
+* 为部分方块和物品增加材质贴图 added textures to some blocks and items
+* 增加方块和物品 added blocks and items
+  * 闪电熔岩 fulgurite block
+  * 末影龙鳞 ender dragon squama item
+  * 龙钢锭 dragonium ingot item
+  * 龙钢块 dragonium block
+  * 艾瑞希亚锭 irisia ingot item
+  * 艾瑞希亚块 irisia block
+* 增加液体 added fluids
+  * 熔融艾瑞希亚 molten irisia
+* 增加材料 added materials
+  * 闪电熔岩 fulgurite
+  * 末影龙鳞 ender dragon squama
+  * 龙钢 dragonium
+  * 艾瑞西亚 irisia
+* 增加属性 added traits
+  * 雷鸣 thundering
+
 ## 0.2.28.0 _2020-01-16 11:00_
 
 * 增加部分合成表 added some recipes